requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.0
jinja2==3.1.6
pydantic==2.11.3
pydantic-settings==2.8.1
grpcio==1.80.0
grpcio-tools==1.80.0
protobuf==6.31.1
pytest==8.3.5
httpx==0.28.1
